% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Categorize quantitative variable into categories based on deciles}
\usage{
categorize(dataset = all_cities, varname)
}
\arguments{
\item{dataset}{a dataset, defaults to all_cities}

\item{varname}{a string corresponding to one of the variables in all_cities.}
}
\value{
a categorized vector
}
\description{
Categorize quantitative variable into categories based on deciles
}
\examples{
categorize(all_cities,"X2018")
}
