% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_glourb.R
\name{connect_to_glourb}
\alias{connect_to_glourb}
\title{Connects to the sandbox database. The necessary parameters are in the R user environment (usethis::edit_r_environ())}
\usage{
connect_to_glourb()
}
\description{
Connects to the sandbox database. The necessary parameters are in the R user environment (usethis::edit_r_environ())
}
