% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_sandbox.R
\name{connect_to_sandbox}
\alias{connect_to_sandbox}
\title{Connects to the sandbox database. The necessary parameters are in the R user environment (usethis::edit_r_environ())}
\usage{
connect_to_sandbox()
}
\description{
Connects to the sandbox database. The necessary parameters are in the R user environment (usethis::edit_r_environ())
}
