% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_city.R
\name{describe_city}
\alias{describe_city}
\title{Describe all characteristics of a city}
\usage{
describe_city(dataset = all_cities, city)
}
\arguments{
\item{dataset}{a dataset, defaults to all_cities}

\item{city}{city name}
}
\value{
a plot
}
\description{
Describe all characteristics of a city
}
\examples{
describe_city(all_cities,"Cordoba--Spain")
all_cities_clust=all_cities \%>\% run_hclust(15)
describe_city(all_cities_clust,"Cordoba--Spain")
}
