% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_clusters.R
\name{describe_clusters}
\alias{describe_clusters}
\title{Title}
\usage{
describe_clusters(dataset, display_ranks = FALSE)
}
\arguments{
\item{display_ranks}{whether to show raw values of quantitative variables or ranks (as calculated by function norm_data) defaults to FALSE}

\item{a}{glourbi dataset with varnames comparable to all_cities, with a cluster descriptor}
}
\value{
a plot gathering boxplots as facets
}
\description{
Description
}
\examples{
data(all_cities)
all_cities_clust=run_hclust(all_cities, nclust=2)
describe_clusters(all_cities_clust,display_ranks=TRUE)
describe_clusters(all_cities_clust,display_ranks=FALSE)
}
