% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city_data.R
\name{get_city_sf}
\alias{get_city_sf}
\title{Get sf data}
\usage{
get_city_sf(name, thisCityCode, conn)
}
\arguments{
\item{name}{the name of the table (with geometry)}

\item{thisCityCode, }{the city code}

\item{conn}{the connection}
}
\value{
the contents of table with name for rows about citycode
}
\description{
Get sf data
}
