% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city_data.R
\name{get_city_tib}
\alias{get_city_tib}
\title{Get tibble data}
\usage{
get_city_tib(name, thisCityCode, conn)
}
\arguments{
\item{name}{the name of the table (no geometry)}

\item{thisCityCode, }{the city code}

\item{conn}{the connection}
}
\value{
the contents of table with name for rows about citycode
}
\description{
Get tibble data
}
