% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color.R
\name{get_color}
\alias{get_color}
\title{Returns the color associated with a certain value of a categorical variable}
\usage{
get_color(varname, value)
}
\arguments{
\item{varname}{name of the variable considered, as string}

\item{value}{name of the value, as string
Description}
}
\value{
a hex color value
}
\description{
Returns the color associated with a certain value of a categorical variable
}
\examples{
get_color("biom","b08")
}
