% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_utils.R
\name{get_txt_segment}
\alias{get_txt_segment}
\title{Returns txt_page data for given citycode and river}
\usage{
get_txt_segment(thisCityCode, thisRiver, conn)
}
\arguments{
\item{thisCityCode}{the citycode}

\item{thisRiver}{the river}

\item{conn}{the connection to glourb database}
}
\value{
a table
}
\description{
Returns txt_page data for given citycode and river
}
