% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_map.R
\name{global_map}
\alias{global_map}
\title{global_map Title}
\usage{
global_map(
  dataset = all_cities,
  varname = "X2018",
  selection1 = "selection1_GSW"
)
}
\arguments{
\item{dataset}{a dataset, defaults to all_cities}

\item{varname}{a string corresponding to one of the variables in all_cities. Defaults to X2018}

\item{selection1}{which selection1 to consider. Can be either selection1_GSW or selection1_Discourses. Defaults to selection1_GSW.}
}
\value{
a leaflet map
}
\description{
global_map Title
}
\examples{
global_map(all_cities,"X2018")
global_map(all_cities,"clim")
}
