% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glourbi-package.R
\docType{package}
\name{glourbi-package}
\alias{glourbi}
\alias{glourbi-package}
\title{glourbi: Glourbi: A Package To Support The GloUrb ANR Project Results}
\description{
This package gathers functions to display the GloUrb results.
}
\author{
\strong{Maintainer}: Lise Vaudor \email{lise.vaudor@ens-lyon.fr} (\href{https://orcid.org/0000-0003-1844-6425}{ORCID})

}
\keyword{internal}
