% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\name{norm_data}
\alias{norm_data}
\title{Normalize columns of a dataset (in case they are numeric)}
\usage{
norm_data(dataset, label = "name", type = "rank")
}
\arguments{
\item{dataset}{a dataset}

\item{label}{the name of the variable used to label individuals. Defaults to "name"}

\item{type}{the type of normalization. Can be either "center_and_scale" or "rank". Defaults to "rank"}
}
\value{
a normalized dataset in which all numeric variables have been centered and scaled
}
\description{
Normalize columns of a dataset (in case they are numeric)
}
\examples{
norm_data(all_cities)
}
