% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distrib.R
\name{plot_distrib}
\alias{plot_distrib}
\title{Plot the distribution of a variable in dataset}
\usage{
plot_distrib(dataset = all_cities, varname, byclass = FALSE)
}
\arguments{
\item{dataset}{a dataset, defaults to all_cities}

\item{varname}{a variable name}

\item{byclass}{whether to plot distributions according to class or not. Defaults to FALSE}
}
\value{
a plot showing the distribution
}
\description{
Plot the distribution of a variable in dataset
}
\examples{
plot_distrib(all_cities,"clim")
all_cities_clust=run_hclust(all_cities)
plot_distrib(all_cities_clust,"X2018")
plot_distrib(all_cities_clust,"X2018",byclass=TRUE)
}
