% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palette.R
\name{plot_palette}
\alias{plot_palette}
\title{Plots the legend showing values for a color palette}
\usage{
plot_palette(dataset = all_cities, varname)
}
\arguments{
\item{dataset}{a dataset, defaults to all_cities}

\item{varname}{a string corresponding to one of the variables in all_cities.}
}
\value{
a plot
}
\description{
Plots the legend showing values for a color palette
}
\examples{
data(all_cities)
plot_palette(all_cities,"clim")
plot_palette(all_cities,"X2018")
}
