% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hclust.R
\name{run_hclust}
\alias{run_hclust}
\title{Runs a hierarchical clustering on dataset}
\usage{
run_hclust(dataset, nclust = 5, namevar = "cluster")
}
\description{
Runs a hierarchical clustering on dataset
}
\examples{
data(all_cities)
all_cities_clust=run_hclust(all_cities)
}
