% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{run_pca}
\alias{run_pca}
\title{Runs a PCA on dataset}
\usage{
run_pca(dataset, quali.sup = NULL)
}
\value{
a PCA object
}
\description{
Runs a PCA on dataset
}
\examples{
data(all_cities)
mypca=run_pca(all_cities)
all_cities_clust=run_hclust(all_cities)
mypca=run_pca(all_cities_clust, quali.sup="cluster")
}
