% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_data.R
\name{sep_data}
\alias{sep_data}
\title{A function that separates variables in the dataset into 3 groups: identifying (vars_id), descriptive (vars_des), numeric (vars_num)}
\usage{
sep_data(dataset)
}
\arguments{
\item{dataset}{a dataset with varnames comparable to all_cities}
}
\value{
a list with elements vars_id, vars_des, vars_cat, vars_num
}
\description{
A function that separates variables in the dataset into 3 groups: identifying (vars_id), descriptive (vars_des), numeric (vars_num)
}
\examples{
sep_data(all_cities)
}
