% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_data.R, R/sep_vars.R
\name{sep_vars}
\alias{sep_vars}
\title{A function that separates variables in the dataset into 3 groups: identifying (vars_id), descriptive (vars_des), numeric (vars_num)}
\usage{
sep_vars(dataset)

sep_vars(dataset)
}
\arguments{
\item{dataset}{a glourbi dataset with varnames comparable to all_cities}
}
\value{
a list with elements vars_id, vars_des, vars_cat, vars_num

a list with elements vars_id, vars_des, vars_cat, vars_num
}
\description{
A function that separates variables in the dataset into 3 groups: identifying (vars_id), descriptive (vars_des), numeric (vars_num)

A function that separates variables in the dataset into 3 groups: identifying (vars_id), descriptive (vars_des), numeric (vars_num)
}
\examples{
sep_vars(all_cities)
sep_vars(all_cities)
}
