% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citycode_or_cityname.R
\name{to_citycode}
\alias{to_citycode}
\title{Returns city code associated to city name}
\usage{
to_citycode(cityname)
}
\arguments{
\item{cityname}{the cityname}
}
\value{
citycode
}
\description{
Returns city code associated to city name
}
\examples{
to_citycode("Denver-Aurora")
}
