% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citycode_or_cityname.R
\name{to_cityname}
\alias{to_cityname}
\title{Returns city name associated to city code}
\usage{
to_cityname(citycode)
}
\arguments{
\item{citycode}{the citycode}
}
\value{
cityname
}
\description{
Returns city name associated to city code
}
\examples{
to_cityname("916904_22567")
}
