% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_map.R
\name{global_map}
\alias{global_map}
\title{global_map Title}
\usage{
global_map(varname = "X2018")
}
\arguments{
\item{varname}{a string corresponding to one of the variables in all_cities. Defaults to X2018}
}
\value{
a leaflet map
}
\description{
global_map Title
}
\examples{
global_map("X2018")
}
