% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale numerical values so that they are centered around 0 and fall between -maxval and maxval}
\usage{
rescale(x, maxval = 1)
}
\description{
Rescale numerical values so that they are centered around 0 and fall between -maxval and maxval
}
\examples{
rescale(rnorm(100,5,3),maxval=10)
}
